Move cf2pulumi to pkg
